Refactor: reorganize test files to be co-located with implementation
